# Another SmallVec.extend optimization