--- a/src/librustc_data_structures/small_vec.rs
+++ b/src/librustc_data_structures/small_vec.rs
@@ -169,18 +169,11 @@ impl<A: Array> FromIterator<A::Element> for SmallVec<A> {
 
 impl<A: Array> Extend<A::Element> for SmallVec<A> {
     fn extend<I: IntoIterator<Item=A::Element>>(&mut self, iter: I) {
-        if self.is_array() {
-            let iter = iter.into_iter();
-            self.reserve(iter.size_hint().0);
-
-            for el in iter {
-                self.push(el);
-            }
-        } else {
-            match self.0 {
-                AccumulateVec::Heap(ref mut vec) => vec.extend(iter),
-                _ => unreachable!()
-            }
+        let iter = iter.into_iter();
+        self.reserve(iter.size_hint().0);
+        match self.0 {
+            AccumulateVec::Heap(ref mut vec) => vec.extend(iter),
+            _ => iter.for_each(|el| self.push(el))
         }
     }
 }