@llogiq 2018-08-01
r? @Mark-Simulacrum