@Manishearth 2018-08-05
Perhaps we should also add this impl to https://github.com/servo/rust-smallvec ?