@Manishearth 2018-08-05
Come to think of it -- is there any reason why we can't unify the smallvecs?